show a buffer for background jobs only when needed

+++ buffers.c
@@ -53,8 +53,11 @@ static int ask_for_closing_and_maybe_terminate(buffer_t *buffer, GtkWidget *wind
 
 	switch (result) {
 	case KILL_AND_CLOSE_RESPONSE:
-		kill(buffer->job->child_pid, SIGTERM);
-		buffer->job->buffer = NULL;
+		if (buffer->job != NULL) {
+			kill(buffer->job->child_pid, SIGTERM);
+			buffer->job->terminating = true;
+			buffer->job->buffer = NULL;
+		}
 		break;
 	case CANCEL_ACTION_RESPONSE:
 	default: return 0;
